refactor : IDE-75 Loading 수정, 폴더 이동
