--- src/app/Loading.module.css
@@ -1,14 +1,0 @@
-.loading {
-  position: absolute;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  place-content: center;
-  place-items: center;
-}
-
-.loading img {
-  width: 64px;
-  height: 64px;
-  opacity: 0.2;
-}
