+++ src/app/ide/Editor/Avatars/Avatars.styles.ts
