+++ src/app/ide/Editor/Avatars/Avatars.tsx
