+++ src/app/ide/Editor/EditorTab/EditorTab.styles.ts
