+++ src/app/ide/page.tsx
@@ -2,7 +2,7 @@
 import React, { useEffect, useRef, useState } from 'react';
 import Chatting from './Chatting/Chatting';
 import { Client } from '@stomp/stompjs';
-import initializeWebSocket from '@/app/api/websocket';
+// import initializeWebSocket from '@/app/api/websocket';
 import useGeneralChatStore, {
   GeneralMessageType,
 } from '@/store/useChattingStore';
@@ -18,7 +18,7 @@ import FileTree from './FileTree/FileTree';
 import TerminalTest from './Terminal/TerminalTest';
 import Toolbar from './Toolbar/Toolbar';
 import { testWebsocket } from '@/app/api/websocket';
-import EditorTab from './Editor/EditorTab';
+import EditorTab from './Editor/EditorTab/EditorTab';
 import ShowEditor from './Editor/ShowEditor';
 import { useFileStore } from '@/store/useFileStore';
 interface ReceivedMessageType {
@@ -37,7 +37,7 @@ const Ide = () => {
 
   useEffect(() => {
     if (clientRef.current == null) {
-      clientRef.current = initializeWebSocket();
+      // clientRef.current = initializeWebSocket();
     }
 
     const client = clientRef.current;
